Enforce POST requests must have json content type, or route override
Want to ensure that we reject post requests to our backend if they have
a type different from application/json. This was attempted in the past
in https://github.com/metabase/metabase/pull/9547/files and then undone
in https://github.com/metabase/metabase/pull/9559/files.

The issue was that some of our query endpoints and download routes take
form params, so we can't just do a blanket "if POST, require
application/json".

But the routes like this are few and far between. The solution:

```clojure
(api/defendpoint ^:streaming ^{:content-types #{:content/json :content/form}} POST
  "/:card-id/query/:export-format"
  ...)
```

This specifies that this route for `POST api/card/<card-id>/query/csv`
can accept _either_ form or json headers. There are a few tests that
might still send a json request so we can leave it for now.

How it works:

Compojure depends on a library called clout. This is the actual routing
bit in a protocol

```clojure
(defprotocol Route
  (route-matches [route request]
    "If the route matches the supplied request, the matched keywords are
    returned as a map. Otherwise, nil is returned."))
```

So for posts, we allow it to construct that implementation, but then we
wrap it. The wrapper checks if it matches, and then if so, checks if the
content-type headers are appropriate.

The matching is based on regexes. Matches a grep-friendly
keyword (`:content/json`, `:content/form`, rather than trying to find
instances of `:json` or `:form` in the codebase).

```clojure
(def ^:private content-type->regex
  "A map of content type to a regex that should match that content type."
  {:content/json #"^application/(.+\+)?json"
   :content/form #"(^application/x-www-form-urlencoded)|(multipart/form-data)|(application/form)"})
```